+++ extensions/vscode/src/multiStepInputs/newCredential.ts
@@ -14,11 +14,12 @@ import {
   getMessageFromError,
   getSummaryStringFromError,
 } from "src/utils/errors";
-import { formatURL, normalizeURL } from "src/utils/url";
+import { formatURL } from "src/utils/url";
 import { checkSyntaxApiKey } from "src/utils/apiKeys";
 import { showProgress } from "src/utils/progress";
 import { openConfigurationCommand } from "src/commands";
 import { extensionSettings } from "src/extension";
+import { findExistingCredentialByURL } from "src/multiStepInputs/common";
 
 const createNewCredentialLabel = "Create a New Credential";
 
@@ -75,10 +76,23 @@ export async function newCredential(
   // ***************************************************************
   async function inputServerUrl(input: MultiStepInput, state: MultiStepState) {
     const thisStepNumber = assignStep(state, "inputServerUrl");
-    const currentURL =
+    let currentURL =
       typeof state.data.url === "string" && state.data.url.length
         ? state.data.url
         : "";
+
+    if (currentURL === "") {
+      currentURL = await extensionSettings.defaultConnectServer();
+    }
+
+    // Two credentials for the same URL is not allowed so clear the default if one is found
+    if (
+      currentURL !== "" &&
+      findExistingCredentialByURL(credentials, currentURL)
+    ) {
+      currentURL = "";
+    }
+
     const url = await input.showInputBox({
       title: state.title,
       step: thisStepNumber,
@@ -112,11 +126,10 @@ export async function newCredential(
             severity: InputBoxValidationSeverity.Error,
           });
         }
-        const existingCredential = credentials.find((credential) => {
-          const existing = normalizeURL(credential.url).toLowerCase();
-          const newURL = normalizeURL(input).toLowerCase();
-          return newURL.includes(existing);
-        });
+        const existingCredential = findExistingCredentialByURL(
+          credentials,
+          input,
+        );
         if (existingCredential) {
           return Promise.resolve({
             message: `Error: Invalid URL (this server URL is already assigned to your credential "${existingCredential.name}". Only one credential per unique URL is allowed).`,