+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Improved the extension's behavior in VS Code workspace restricted mode.
+  The extension will activate, but prompt to manage workspace trust so it can initialize.
+
 ### Fixed
 
 ### Changed