+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a new config setting `positPublisher.defaultConnectServer` to pre-populate the Server URL. Also checks the [default RStudio/Posit Workbench setting](https://docs.posit.co/ide/server-pro/rstudio_pro_sessions/rstudio_connect_server.html) (rstudio-pro#7466)
+
 - Improved the extension's behavior in VS Code workspace restricted mode.
-  The extension will activate, but prompt to manage workspace trust so it can initialize.
+  The extension will activate, but prompt to manage workspace trust so it can initialize. (#2611)
 
 ### Fixed