+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.1]
+
+### Fixed
+
+- Fix missing platform build in OpenVSX due to outage.
+
+### Changed
+
+- Updated `record.toml` to correspond with updated schema change in v1.12
+
 ## [1.12.0]
 
 ### Added